move load_balancing to new execution mod
