+++ scylla/src/transport/iterator.rs
@@ -25,6 +25,7 @@ use crate::session::RequestSpan;
 use crate::cluster::ClusterData;
 use crate::cluster::{Node, NodeRef};
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
+use crate::execution::load_balancing::{self, RoutingInfo};
 use crate::frame::{
     response::{
         result,
@@ -35,7 +36,6 @@ use crate::frame::{
 use crate::history::{self, HistoryListener};
 use crate::statement::Consistency;
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
-use crate::transport::load_balancing::{self, RoutingInfo};
 use crate::transport::metrics::Metrics;
 use crate::transport::retry_policy::{QueryInfo, RetryDecision, RetrySession};
 use tracing::{trace, trace_span, warn, Instrument};
